backend/requirements.txt
# Aerogate Backend Dependencies
# Complete production-ready requirements file

# ========================================
# Web Framework & Server
# ========================================
fastapi>=0.115.0
uvicorn>=0.32.0
gunicorn>=21.0.0
mangum>=0.17.0

# ========================================
# Database & ORM
# ========================================
sqlmodel>=0.0.22
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
aiosqlite>=0.19.0
pgvector>=0.2.0

# ========================================
# Pydantic & Validation
# ========================================
pydantic>=2.0.0
pydantic-settings>=2.6.0
email-validator>=2.0.0

# ========================================
# Authentication & Security
# ========================================
passlib>=1.7.4
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# ========================================
# Configuration & Environment
# ========================================
python-dotenv>=1.0.1

# ========================================
# Logging
# ========================================
loguru>=0.7.3

# ========================================
# File Handling
# ========================================
python-multipart>=0.0.17

# ========================================
# ML / Embeddings (CPU-only for smaller Docker image)
# ========================================
# This special index provides CPU-only PyTorch (~400MB vs 2GB+ with CUDA)
# This special index provides CPU-only PyTorch (~400MB vs 2GB+ with CUDA)
# torch==2.1.0
# sentence-transformers>=2.2.0
# numpy>=1.24.0
